+++ src/app/shared/settings-service/settings.service.spec.ts
@@ -1,19 +1,19 @@
 import { TestBed } from '@angular/core/testing';
 import { SettingsService } from './settings.service';
-import { FhirBackendService } from '../fhir-backend/fhir-backend.service';
-import { FhirBackendModule } from '../fhir-backend/fhir-backend.module';
-import { RouterTestingModule } from '@angular/router/testing';
+import { configureTestingModule } from 'src/test/helpers';
+import { HttpTestingController } from '@angular/common/http/testing';
 
 describe('SettingsService', () => {
   let service: SettingsService;
-  let fhirBackendService: FhirBackendService;
+  let mockHttp: HttpTestingController;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [FhirBackendModule, RouterTestingModule]
-    });
-    fhirBackendService = TestBed.inject(FhirBackendService);
+  beforeEach(async () => {
+    await configureTestingModule(
+      {},
+      { serverUrl: 'https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1' }
+    );
     service = TestBed.inject(SettingsService);
+    mockHttp = TestBed.inject(HttpTestingController);
   });
 
   it('should be created', () => {
@@ -21,10 +21,6 @@ describe('SettingsService', () => {
   });
 
   it('should read custom settings for dbGap', async () => {
-    spyOnProperty(fhirBackendService, 'serviceBaseUrl').and.returnValue(
-      'https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1'
-    );
-    await service.loadJsonConfig().toPromise();
     expect(typeof service.get('definitionsFile')).toBe('string');
   });
 });